requirements.txt
bentoml==1.4.15
catboost==1.2.8
category_encoders==2.8.1
eli5==0.16.0
fastapi==0.115.12
joblib==1.4.2
matplotlib==3.10.3
missingno==0.5.2
mysql_connector_repackaged==0.3.1
numpy==2.3.0
pandas==2.3.0
prometheus_client==0.21.1
prometheus_fastapi_instrumentator==7.1.0
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.4
scikit_learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
SQLAlchemy==1.4.54
